move *Info structs out of the hdf5 file
